Rename `domain` to `core`, trying to be a touch more idiomatic
