+++ src/sudoku/core/consts.rs
