+++ src/sudoku/core/validatable_units.rs
