+++ src/sudoku/mod.rs
@@ -1,2 +1,2 @@
-pub mod domain;
+pub mod core;
 pub mod format;
